backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pandas==2.1.1
google-generativeai==0.3.2
python-dotenv==1.0.0
pymongo==4.6.1
waitress==2.1.2